Title: 코드트리 DB, Time: 2022ms, Memory: 22MB, Status: Time Limit Exceed - Codetree

+++ 250331/코드트리 DB/codetree-db.cpp
@@ -14,7 +14,7 @@ struct Leaf {
 struct Node {
     int min;
     int max;
-    int sum;
+    long long sum;
     Node* left;
     Node* right;
     Node* parents;
